+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > Add extention class to extend the session object for simplified manipulation of the server's registered users
 
 * **Add** - `UserUtils.cs`
+* **Add** - 'GuaranteedRate.Examples.UserUtils' for `UserUtils.cs`
 
 ## v1.7.1 / 2015 Nov 19